# smarthire-backend/specs/001-smarthire-backend-platform/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: SmartHire Backend API
  description: |
    Internal recruitment management platform API.
    All endpoints require a valid JWT bearer token except `/prescreen/**`, `/health`, `/docs`, and `/openapi.json`.
  version: "1.0.0"
  contact:
    name: SmartHire Platform Team
servers:
  - url: "http://localhost:8083"
    description: Local development
  - url: "https://smarthire.internal"
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResponseDto:
      type: object
      properties:
        response:
          type: array
          items: {}
        message:
          type: string
        exception:
          type: string
          nullable: true
        result:
          type: string
          nullable: true
      required:
        - response
        - message
...
    FreeSlotRequest:
      type: object
      required: [interviewerId, interviewDate, startTime, interviewTypeId, technologyId]
      properties:
        interviewerId:
          type: integer
          format: int64
        interviewDate:
          type: string
          format: date
        startTime:
          type: string
          example: "10:00"
        interviewTypeId:
          type: integer
          format: int64
        technologyId:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        buId:
          type: integer
          format: int64
        participationType:
          type: string
          enum: [Virtual, In-person]
    InterviewSlotRequest:
      type: object
      required: [interviewerCalendarId, candidateId, recruiterId]
      properties:
        interviewerCalendarId:
          type: integer
          format: int64
        candidateId:
          type: integer
          format: int64
        recruiterId:
          type: integer
          format: int64
        interviewTypeId:
          type: integer
          format: int64
    FeedbackRequest:
      type: object
      required: [interviewerCalendarId, feedbackStatus, responseJson]
      properties:
        interviewerCalendarId:
          type: integer
          format: int64
        feedbackStatus:
          type: string
          enum: [Select, Reject, "On Hold", "L1 FB NA", "L2 FB NA", "L3 FB NA"]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        responseJson:
          type: object
          additionalProperties: true
    RescheduleRequest:
      type: object
      required: [interviewerCalendarId, newDate, newStartTime]
      properties:
        interviewerCalendarId:
          type: integer
          format: int64
        newDate:
          type: string
          format: date
        newStartTime:
          type: string
          example: "14:00"

    # -------------------------------------------------------------------------
    # Feedback Form
    # -------------------------------------------------------------------------
...
    LookupRequest:
      type: object
      properties:
        screenId:
          type: string
    MarketUnitRequest:
      type: object
      required: [buId]
      properties:
        buId:
          type: integer
          format: int64

    # -------------------------------------------------------------------------
    # Reports
    # -------------------------------------------------------------------------
...
    KeycloakTokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
        userId:
          type: string
paths:
...
  /interviewer/getInterviewerSlots:
    post:
      tags: [Interviewer]
      summary: Retrieve available interviewer slots by skill, date range, and BU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                technologyId:
                  type: integer
                  format: int64
                fromDate:
                  type: string
                  format: date
                toDate:
                  type: string
                  format: date
                buId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Available interviewer slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /interviewer/saveFreeSlot:
    post:
      tags: [Interviewer]
      summary: Register an interviewer's availability slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreeSlotRequest"
      responses:
        "200":
          description: Free slot saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /interviewer/saveInterviewSlot:
    post:
      tags: [Interviewer]
      summary: Book an interview slot (link candidate to interviewer slot)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewSlotRequest"
      responses:
        "200":
          description: Slot booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /interviewer/saveFeedback:
    post:
      tags: [Interviewer]
      summary: Submit interview feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "200":
          description: Feedback saved; PDF generation triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /interviewer/rescheduledRequest:
    post:
      tags: [Interviewer]
      summary: Reschedule an interview slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleRequest"
      responses:
        "200":
          description: Slot rescheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /interviewer/deleteSlot:
    delete:
      tags: [Interviewer]
      summary: Cancel / delete an interview slot
      parameters:
        - name: interviewerCalendarId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Slot deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Recruiter
  # ===========================================================================
...
  /lookup/fetchDropdown:
    post:
      tags: [Lookup]
      summary: Fetch dropdown values for a given screen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LookupRequest"
      responses:
        "200":
          description: Dropdown items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /lookup/fetchMarketUnit:
    post:
      tags: [Lookup]
      summary: Fetch market units for a BU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketUnitRequest"
      responses:
        "200":
          description: Market units
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
...
  /keycloak/fetchAllKeycloakUsers:
    post:
      tags: [Keycloak]
      summary: Fetch all users from the Keycloak realm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeycloakTokenRequest"
      responses:
        "200":
          description: Keycloak user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /keycloak/deleteEmployee:
    delete:
      tags: [Keycloak]
      summary: Delete an employee from local DB and Keycloak
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /keycloak/updateEmployee:
    put:
      tags: [Keycloak]
      summary: Push employee updates to Keycloak
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeycloakTokenRequest"
      responses:
        "200":
          description: Employee updated in Keycloak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Teams Meeting
  # ===========================================================================
...
  /lookup/{entity}:
    get:
      tags: [Lookup]
      summary: Generic lookup by entity name
      parameters:
        - name: entity
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lookup results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Prescreen / Referral (app-specific paths)
  # ===========================================================================
...